requirements.txt
streamlit
python-dotenv
PyPDF2
langchain-community
langchain
openai
faiss-cpu
altair
tiktoken
huggingface_hub
pyth
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2